refactor: util 의존성 개선 / 패키지 구조 변경
