+++ commons/common-util/src/main/kotlin/kr/dataportal/calendar/util/hashExtensions.kt
@@ -1,4 +1,4 @@
-package kr.dataportal.calendar
+package kr.dataportal.calendar.util
 
 import java.security.MessageDigest
