+++ commons/common-util/src/main/kotlin/kr/dataportal/calendar/util/jsonExtensions.kt
@@ -1,4 +1,4 @@
-package kr.dataportal.calendar
+package kr.dataportal.calendar.util
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect
 import com.fasterxml.jackson.annotation.PropertyAccessor
