+++ commons/common-util/src/main/kotlin/kr/dataportal/calendar/util/timeExtensions.kt
@@ -1,4 +1,4 @@
-package kr.dataportal.calendar
+package kr.dataportal.calendar.util
 
 import java.time.Instant
 import java.time.LocalDateTime
