+++ consortium-calendar-assets/consortium-calendar-assets-domain/build.gradle.kts
@@ -1,3 +1,5 @@
 dependencies {
+    implementation(project(":commons:common-model"))
+    implementation(project(":commons:common-util"))
     implementation(project(":consortium-calendar-assets:consortium-calendar-assets-infrastructure:persistence-database"))
 }
