+++ consortium-calendar-assets/consortium-calendar-assets-domain/src/test/kotlin/kr/dataportal/calendar/account/service/RegisterPlainAuthAccountTest.kt
@@ -7,11 +7,11 @@ import io.mockk.mockk
 import io.mockk.slot
 import kr.dataportal.calendar.account.exception.AlreadyRegisteredEmailException
 import kr.dataportal.calendar.account.usecase.RegisterPlainAuthAccountUseCase
-import kr.dataportal.calendar.hashSHA512
 import kr.dataportal.calendar.persistence.entity.account.AccountAuthenticationJpaEntity
 import kr.dataportal.calendar.persistence.entity.account.AccountJpaEntity
 import kr.dataportal.calendar.persistence.repository.account.AccountAuthenticationRepository
 import kr.dataportal.calendar.persistence.repository.account.AccountRepository
+import kr.dataportal.calendar.util.hashSHA512
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
